--- wanted-cqrs/src/main/java/com/mkhwang/wantedcqrs/product/domain/dto/ReviewDto.java
@@ -1,4 +1,0 @@
-package com.mkhwang.wantedcqrs.product.domain.dto;
-
-public class ReviewDto {
-}
