+++ wanted-cqrs/src/main/java/com/mkhwang/wantedcqrs/product/domain/dto/category/CategorySearchDto.java
@@ -1,4 +1,4 @@
-package com.mkhwang.wantedcqrs.product.domain.dto;
+package com.mkhwang.wantedcqrs.product.domain.dto.category;
 
 import com.mkhwang.wantedcqrs.common.dto.PageRequestDto;
 import lombok.Getter;
