+++ wanted-cqrs/src/main/java/com/mkhwang/wantedcqrs/product/domain/dto/category/CategorySearchResultDto.java
@@ -1,4 +1,4 @@
-package com.mkhwang.wantedcqrs.product.domain.dto;
+package com.mkhwang.wantedcqrs.product.domain.dto.category;
 
 import com.mkhwang.wantedcqrs.config.advice.dto.ApiPageInfo;
 import lombok.Getter;
