+++ wanted-cqrs/src/main/java/com/mkhwang/wantedcqrs/product/domain/dto/review/ReviewDto.java
@@ -1,0 +1,4 @@
+package com.mkhwang.wantedcqrs.product.domain.dto.review;
+
+public class ReviewDto {
+}
