+++ wanted-cqrs/src/main/java/com/mkhwang/wantedcqrs/product/infra/ReviewSearchRepository.java
@@ -1,8 +1,8 @@
 package com.mkhwang.wantedcqrs.product.infra;
 
 import com.mkhwang.wantedcqrs.product.domain.dto.ProductRatingDto;
-import com.mkhwang.wantedcqrs.product.domain.dto.ReviewSearchDto;
-import com.mkhwang.wantedcqrs.product.domain.dto.ReviewSearchResultDto;
+import com.mkhwang.wantedcqrs.product.domain.dto.review.ReviewSearchDto;
+import com.mkhwang.wantedcqrs.product.domain.dto.review.ReviewSearchResultDto;
 
 public interface ReviewSearchRepository {
   ProductRatingDto findAverageRatingByProductId(Long productId);
